+++ diesis/main.py
@@ -4,7 +4,7 @@ from diesis import FileScanner, Config
 def main():
     try:
         Config.Config.setup_from_cli()
-        print('Running Diesis 1.0.4')
+        print('Running Diesis 1.0.5')
         scanner = FileScanner.FileScanner()
         scanner.scan()
         print('All operations completed, bye!')